More work on responsive layout.
For wide screens and 3-column layout, I use a grid with 9 columns to get closer
to the ratios in the AMC proposal. For intermediate screens and 3-column
layout, keep the second sidebar at the bottom of the page but make each block
1/2 the width instead of 1/3.

Also adjust the font size, padding, etc. for block headers in the second
sidebar.